+++ samples/OAuthHttpClientSample/Program.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Threading.Tasks;
@@ -25,6 +26,7 @@ namespace OAuthHttpClientSample
                                hostContext.Configuration["OAuth:TokenId"],
                                hostContext.Configuration["OAuth:TokenSecret"]);
                        options.SignedAsQuery = hostContext.Configuration.GetValue("OAuth:SignedAsQuery", false);
+                       options.SignedAsBody = hostContext.Configuration.GetValue("OAuth:SignedAsBody", false);
                    }).ConfigureHttpClient(client =>
                    {
                        client.DefaultRequestHeaders.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json"));
@@ -36,7 +38,13 @@ namespace OAuthHttpClientSample
             var oauthClient = host.Services.GetRequiredService<OAuthHttpClient>();
 
             Console.WriteLine("Sending a request...");
-            var request = new HttpRequestMessage(HttpMethod.Get, configuration.GetValue<Uri>("OAuth:ResourceUri"));
+            var method = new HttpMethod(configuration.GetValue("Request:Method", HttpMethod.Get.Method));
+            var request = new HttpRequestMessage(method, configuration.GetValue<Uri>("Request:Uri"));
+            var body = configuration.GetSection("Request:Body").Get<IDictionary<string, string>>();
+            if (body != null)
+            {
+                request.Content = new FormUrlEncodedContent(body);
+            }
             var response = await oauthClient.HttpClient.SendAsync(request).ConfigureAwait(false);
             var data = await response.Content.ReadAsStringAsync().ConfigureAwait(false);
             Console.WriteLine("Response data:");