+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## 2020-07-26
 
+
+### GSS.Authorization.OAuth.HttpClient 1.3.1
+
+- Fixed null Request.Content during SignedAsBody
+
 ### GSS.Authorization.OAuth.HttpClient 1.3.0
 
 - Support to signed as form body